seperate functions out and refactoring
