+++ NFACT/preprocess/nfactpp.py
@@ -1,21 +1,10 @@
-import os
-import re
-import shutil
-
 # NFACT functions
-from NFACT.base.imagehandling import get_file
-from NFACT.base.utils import colours, error_and_exit
 from .nfactpp_setup import (
-    check_surface_arguments,
+    check_seeds_surfaces,
     nfact_pp_folder_setup,
     check_roi_seed_len,
 )
-from .nfactpp_functions import (
-    hcp_get_seeds,
-    hcp_get_rois,
-    hcp_reorder_seeds_rois,
-    update_seeds_file,
-)
+from .nfactpp_functions import update_seeds_file, get_file
 from .probtrackx_functions import (
     build_probtrackx2_arguments,
     write_options_to_file,
@@ -23,6 +12,11 @@ from .probtrackx_functions import (
     get_target2,
     seeds_to_ascii,
 )
+from NFACT.base.utils import colours, error_and_exit
+
+import os
+import re
+import shutil
 
 
 def pre_processing(arg: dict, handler) -> None:
@@ -40,7 +34,7 @@ def pre_processing(arg: dict, handler) -> None:
     None
     """
 
-    arg["surface"] = check_surface_arguments(arg["seed"], arg["rois"])
+    arg["surface"] = check_seeds_surfaces(arg["seed"], arg["rois"])
     col = colours()
 
     if arg["surface"]:
@@ -134,79 +128,3 @@ def pre_processing(arg: dict, handler) -> None:
             update_seeds_file(os.path.join(sub, arg["outdir"], "seeds.txt"))
             for sub in arg["list_of_subjects"]
         ]
-
-
-def hcp_stream_main(arg: dict, handler: object) -> None:
-    """
-    hcp stream main function
-
-    Parameters
-    ----------
-    arg: dict
-       dictionary of command line
-       arguments
-
-    Returns
-    ------
-    None
-
-    """
-    col = colours()
-    print(f'{col["darker_pink"]}HCP stream selected{col["reset"]}')
-    subjects_commands = []
-    print("Number of subjects: ", len(arg["list_of_subjects"]))
-    for sub in arg["list_of_subjects"]:
-        # looping over subjects and building out directories
-        print(f"\n{col['pink']}Setting up:{col['reset']} {os.path.basename(sub)}")
-        seeds = hcp_get_seeds(sub)
-        arg["rois"] = hcp_get_rois(sub)
-        nfactpp_diretory = os.path.join(sub, arg["outdir"])
-
-        if arg["overwrite"]:
-            if os.path.exists(nfactpp_diretory):
-                print(
-                    f'{col["red"]}{arg["outdir"]} directory already exists. Overwriting{col["reset"]}'
-                )
-                shutil.rmtree(nfactpp_diretory, ignore_errors=True)
-
-        nfact_pp_folder_setup(nfactpp_diretory)
-
-        ordered_by_hemisphere = hcp_reorder_seeds_rois(seeds, arg["rois"])
-        for hemishphere, img in ordered_by_hemisphere.items():
-            seeds_to_ascii(
-                img[0],
-                img[1],
-                os.path.join(
-                    nfactpp_diretory, "files", f"{hemishphere}_white.32k_fs_LR.surf.asc"
-                ),
-            )
-
-        asc_seeds = [
-            os.path.join(nfactpp_diretory, "files", "left_white.32k_fs_LR.surf.asc"),
-            os.path.join(nfactpp_diretory, "files", "right_white.32k_fs_LR.surf.asc"),
-        ]
-        seed_text = "\n".join(asc_seeds)
-        error_and_exit(write_options_to_file(nfactpp_diretory, seed_text))
-
-        if not arg["target2"]:
-            print(
-                f'{col["purple"]}No target given. Creating a whole brain target.{col["reset"]}'
-            )
-
-            get_target2(
-                arg["ref"],
-                os.path.join(nfactpp_diretory, "files", "target2"),
-                arg["res"],
-                arg["ref"],
-                "nearestneighbour",
-            )
-        subjects_commands.append(build_probtrackx2_arguments(arg, sub, hcp_stream=True))
-
-    if arg["n_cores"]:
-        handler.set_suppress_messages = True
-
-    Probtrackx(subjects_commands, arg["cluster"], arg["n_cores"])
-    [
-        update_seeds_file(os.path.join(sub, arg["outdir"], "seeds.txt"))
-        for sub in arg["list_of_subjects"]
-    ]
