+++ NFACT/preprocess/nfactpp_functions.py
@@ -1,4 +1,6 @@
 from NFACT.base.utils import error_and_exit
+from NFACT.base.imagehandling import check_files_are_imaging_files
+import os
 
 
 def update_seeds_file(file_path: str) -> None:
@@ -24,3 +26,34 @@ def update_seeds_file(file_path: str) -> None:
             file.write(update_extensions)
     except Exception as e:
         error_and_exit(False, f"Unable to change seeds file due to {e}")
+
+
+# TODO: seperate this function out
+def get_file(img_file: list, sub: str) -> list:
+    """
+    Function to get an imaging file
+    type and returns it. Checks that file
+    is correct file type and exists.
+
+    Parameters
+    ----------
+    img_file: list
+        a list of imaging files
+    sub: str
+       path to subjects directory.
+
+    Returns
+    -------
+    img_files: list
+        list of imging files
+
+    """
+    img_files = [os.path.join(sub, file) for file in img_file]
+    [
+        error_and_exit(
+            os.path.exists(path), f"Unable to find {path}. Please check it exists"
+        )
+        for path in img_files
+    ]
+    [check_files_are_imaging_files(path) for path in img_files]
+    return img_files
